chore(password-input): move to packages
