+++ packages/components/inputs/password-input/src/index.js
@@ -1,0 +1,3 @@
+export { default } from './password-input';
+
+export { default as version } from './version';
