+++ packages/components/inputs/password-input/src/password-input.visualspec.js
