--- src/components/inputs/password-input/index.js
@@ -1,1 +1,0 @@
-export { default } from './password-input';
